backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
backend==0.2.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
crispy-forms-gds==0.2.2
cryptography==3.4.7
dateparser==1.0.0
defusedxml==0.7.1
Django==3.0.14
django-crispy-forms==1.12.0
django-email-users==1.1.0
django-filter==2.4.0
django-rest-authtoken==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==8.11.0
flake8==3.8.3
gunicorn==20.0.4
idna==3.2
isort==5.5.1
itypes==1.2.0
Jinja2==3.0.1
lark-parser==0.11.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
ordered-set==4.0.2
pbr==5.6.0
Pillow==8.1.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
recipe==0.26.1
recipes==0.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
SQLAlchemy==1.3.24
sqlparse==0.4.1
stevedore==3.3.0
sureberus==0.14.0
tablib==3.0.0
text-unidecode==1.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
